Dedup responses
Multiple idential RR records should be dedupped so that only one is
left.
I.e.

multiport.th1.kubernetes.local. 3600 IN A   1.1.1.1
multiport.th1.kubernetes.local. 3600 IN A   1.1.1.1

Should return:

multiport.th1.kubernetes.local. 3600 IN A   1.1.1.1

Note: I believe the clients don't really care about this and stuff just
works, but a good server should not be sending out invalid messages.

Also added test case that checks for this. We also dedup the additional
section, but for now there is no test for this.

Fixes #149